feat: add central TypeScript type definitions with JSDoc (#353)
- Add backend/src/types/index.ts with JSDoc for all shared types
- Add frontend/src/types/index.ts with JSDoc for API response types
- Refactor services to import from central types (re-export for compat)
- Update frontend/src/lib/api.ts to re-export types from @/types
- Expand README API Reference with types table and response types
